+++ CHANGELOG.md
@@ -15,7 +15,8 @@
 
 ### Bug fixes
 - Fixed `setup.py` not being able to import `versioneer` when installing in an embedded Python environment. @rly (#662)
-- Fix broken tests in Python 3.10. @rly (#664)
+- Fixed broken tests in Python 3.10. @rly (#664)
+- Fixed broken LaTeX PDF build of the docs. @oruebel (#669)
 
 ## HDMF 3.1.1 (July 29, 2021)